+++ Source/Moka/Moka/GameObjects/interactiveObject.h
@@ -7,26 +7,26 @@
 class InteractiveObject
 {
 public:
-							InteractiveObject(std::string name, std::string type, std::string attachedTo, int x, int y, int width, int height);
+							InteractiveObject(std::string name, std::string type, std::string attachedTo, float x, float y, float width, float height);
 							// ALW - Allow copy construction and assignment
 
 	const std::string &		getName() const;
 	const std::string &		getType() const;
 	const std::string &		getAttachedTo() const;
-	int						getX() const;
-	int						getY() const;
-	int						getWidth() const;
-	int						getHeight() const;
+	float					getX() const;
+	float					getY() const;
+	float					getWidth() const;
+	float					getHeight() const;
 
 
 private:
 	std::string				mName;
 	std::string				mType;
 	std::string				mAttachedTo;
-	int						mX;
-	int						mY;
-	int						mWidth;
-	int						mHeight;
+	float					mX;
+	float					mY;
+	float					mWidth;
+	float					mHeight;
 };
 
 #endif
